RDKB-61398:ccsp-webui:RDKB-59711, RDKB-58524, RDKB-60107, RDKCOM-5316
Reason for change:
A dummy Gerrit was created containing only a commit message for
reference. It links all dependent tickets related to the
ccsp-webui-bci-jst and ccsp-webui-bci-php components, which are being
released with the new tag: 1.0.2.

Weekly RDKCentral Components Release
Date: [22/08/2025] – [29/08/2025]

Test Procedure:Validate the deployed changes mentioned in the ticket
Risks: Low
Priority: P1
Signed-off-by: Goutam_Damodaran@comcast.com
Change-Id: Ie2aa35477e4ba78397316206fffa8cc7d99735a9